break out the checkbox fields component
