+++ app/webpack/reactjs/components/open_studios/open_studios_info_form.tsx
@@ -4,6 +4,7 @@ import { MauCheckboxField } from "@reactjs/components/mau_checkbox_field";
 import { MauTextField } from "@reactjs/components/mau_text_field";
 import { FieldError } from "@reactjs/components/field_error";
 import { OpenStudiosRegistration } from "@reactjs/components/open_studios/open_studios_registration";
+import { SpecialEventScheduleFields } from "@reactjs/components/open_studios/special_event_schedule_fields";
 import { DateTime } from "luxon";
 import { parseTimeSlot } from "@js/app/time_utils";
 import { api } from "@js/services/api";
@@ -17,7 +18,7 @@ interface OpenStudiosInfoFormProps {
   artistId: number;
   participant: types.OpenStudiosParticipant;
   openStudiosEvent: types.OpenStudiosEvent;
-  onUpdateParticipant: (participant: OpenStudiosParticipant) => void;
+  onUpdateParticipant: (participant: types.OpenStudiosParticipant) => void;
 }
 
 // Seems to help with proving the controlled v uncontrolled inputs and Formik
@@ -32,54 +33,7 @@ function denullify(participant) {
   return result;
 }
 
-interface SpecialEventScheduleFieldsProps {
-  specialEvent: types.SpecialEventDetails;
-}
 
-interface TimeSlot {
-  start: string;
-  end: string;
-}
-interface TimeSlotCheckBoxProps {
-  timeslot: TimeSlot;
-  name: string;
-}
-
-const generateSlotFieldName = (timeslot: string): string =>
-  `videoConferenceSchedule[${timeslot}]`;
-
-const formatTimeSlot = (timeslot: types.TimeSlot): string => {
-  const { start, end } = timeslot;
-  return `${start.toLocaleString(DateTime.TIME_SIMPLE)} - ${end.toLocaleString(
-    DateTime.TIME_SIMPLE
-  )} ${start.toLocaleString(DateTime.DATE_MED)} `;
-};
-
-const TimeSlotCheckBox: FC<TimeSlotCheckBoxProps> = ({ timeslot, name }) => {
-  const label = formatTimeSlot(timeslot);
-  return <MauCheckboxField label={label} name={name} />;
-};
-
-const SpecialEventScheduleFields: FC<SpecialEventScheduleFieldsProps> = ({
-  specialEvent,
-}) => {
-  if (!specialEvent?.timeSlots) {
-    return null;
-  }
-  const slots = specialEvent.timeSlots;
-
-  return (
-    <div className="open-studios-info-form__special-event-schedule">
-      {slots.map((slot: string) => {
-        const parsed = parseTimeSlot(slot);
-        const slotName = generateSlotFieldName(slot);
-        return (
-          <TimeSlotCheckBox key={slotName} timeslot={parsed} name={slotName} />
-        );
-      })}
-    </div>
-  );
-};
 
 export const OpenStudiosInfoForm: FC<OpenStudiosInfoFormProps> = ({
   location,
